+++ ddbmock/routes/delete_item.py
@@ -1,17 +1,21 @@
 # -*- coding: utf-8 -*-
 
-from ddbmock.utils import load_table
+from . import load_table
+from ddbmock.utils import push_write_throughput
 
 @load_table
 def delete_item(post, table):
     item = table.delete_item(post[u'Key'], post[u'Expected'])
 
+    capacity = item.get_size().as_units()
+    push_write_throughput(table.name, capacity)
+
     if post[u'ReturnValues'] == "ALL_OLD":
         return {
-            "ConsumedCapacityUnits": item.get_size().as_units(),
+            "ConsumedCapacityUnits": capacity,
             "Attributes": item,
         }
     else:
         return {
-            "ConsumedCapacityUnits": item.get_size().as_units(),
+            "ConsumedCapacityUnits": capacity,
         }