+++ ddbmock/routes/put_item.py
@@ -1,14 +1,17 @@
 # -*- coding: utf-8 -*-
 
-from ddbmock.utils import load_table
+from . import load_table
+from ddbmock.utils import push_write_throughput
 
 @load_table
 def put_item(post, table):
     old, new = table.put(post[u'Item'], post[u'Expected'])
-    units = max(old.get_size().as_units(), new.get_size().as_units())
+    capacity = max(old.get_size().as_units(), new.get_size().as_units())
+
+    push_write_throughput(table.name, capacity)
 
     ret = {
-        "ConsumedCapacityUnits": units,
+        "ConsumedCapacityUnits": capacity,
     }
 
     if post[u'ReturnValues'] == "ALL_OLD":