+++ ddbmock/utils/__init__.py
@@ -1,11 +1,18 @@
 # -*- coding: utf-8 -*-
 
-from ddbmock.database import dynamodb
+from .stat import Stat
 
-def load_table(func):
-    def loader(post, *args):
-        name = post[u'TableName']
-        table = dynamodb.get_table(name)
+perf_logger = {
+    'read': {},
+    'write': {},
+}
 
-        return func(post, table, *args)
-    return loader
+def push_read_throughput(table_name, value):
+    if table_name not in perf_logger['read']:
+        perf_logger['read'][table_name] = Stat("`%s` read throughput" % table_name, 1, 5*60)
+    perf_logger['read'][table_name].push(value)
+
+def push_write_throughput(table_name, value):
+    if table_name not in perf_logger['write']:
+        perf_logger['write'][table_name] = Stat("`%s` write throughput" % table_name, 1, 5*60)
+    perf_logger['write'][table_name].push(value)